Make building firmware from the project easier
This adds a module.json that we can use to build the firmware, as
well as adding a simple section to the readme for building

In order to be 'normal' in yotta terms I've moved the CPP files into the
source directory, however it would also be feasible to add the
module.json at the top level so it built there. I've gone this way to
make it clear exactly what's in the firmware versus the other parts of
the project.